Altered `total_price` column of `orders` table to float

+++ orders/models/orders_models.go
@@ -8,7 +8,7 @@ type Order struct {
 	ID        int    `json:"id"`          // Unique identifier for the order.
 	UserID    string    `json:"user_id"`     // ID of the user who placed the order.
 	AddressID int    `json:"address_id"`  // ID of the address associated with the order.
-	TotalPrice int  `json:"total_price"`  // Total price of the order.
+	TotalPrice float64  `json:"total_price"`  // Total price of the order.
 	CreatedAt time.Time `json:"created_at"`  // Timestamp indicating when the order was created.
 	Status    string `json:"status"`      // Current status of the order.
 }
@@ -46,7 +46,7 @@ type DetailedOrders struct {
 type OrderRequestParams struct {
 	UserID    string    `json:"user_id"`      // ID of the user placing the order.
 	AddressID int    `json:"address_id"`   // ID of the address associated with the order.
-	TotalPrice int  `json:"total_price"`   // Total price of the order.
+	TotalPrice float64  `json:"total_price"`   // Total price of the order.
 	Status    string `json:"status"`       // Current status of the order.
 }
